apps/audio-diarizer/pyproject.toml
[project]
name = "audio-diarizer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch==2.8.0",
    "pyannote.audio",
    "shared-messaging",
    "shared-storage",
    "shared-schemas",
    "pydantic-settings",
]

[tool.uv.sources]
shared-messaging = { workspace = true }
shared-schemas = { workspace = true }

